packfile.c: set shift limit to sizeof(size_t), not long
The code base now uses size_t for all memsized variables.
Allow shift to reach that bitness level.

+++ packfile.c
@@ -1057,7 +1057,7 @@ size_t unpack_object_header_buffer(const unsigned char *buf,
 	size = c & 15;
 	shift = 4;
 	while (c & 0x80) {
-		if (len <= used || bitsizeof(long) <= shift) {
+		if (len <= used || bitsizeof(size_t) <= shift) {
 			error("bad object header");
 			size = used = 0;
 			break;
